publish orderbook to sidecar
